refactor: generics possession getter
