+++ src/main/java/school/hei/patrimoine/patrilang/grammar/PatriLangParser.g4
@@ -68,19 +68,19 @@ compte
     : PUCE TEXT COMMA MOT_VALANT argent date
     ;
 fluxArgentTransferer
-    : PUCE BACKTICK TEXT BACKTICK date COMMA MOT_JE MOT_TRANSFERER argent MOT_DEPUIS TEXT MOT_VERS TEXT
+    : PUCE BACKTICK TEXT BACKTICK date COMMA MOT_JE? MOT_TRANSFERER argent MOT_DEPUIS TEXT MOT_VERS TEXT
     ;
 fluxArgentEntrer
-    : PUCE BACKTICK TEXT BACKTICK date COMMA MOT_JE MOT_ENTRER argent MOT_VERS TEXT
+    : PUCE BACKTICK TEXT BACKTICK date COMMA MOT_JE? MOT_ENTRER argent MOT_VERS TEXT
     ;
 fluxArgentSortir
-    : PUCE BACKTICK TEXT BACKTICK date COMMA MOT_JE MOT_SORTIR argent MOT_DEPUIS TEXT
+    : PUCE BACKTICK TEXT BACKTICK date COMMA MOT_JE? MOT_SORTIR argent MOT_DEPUIS TEXT
     ;
 acheterMateriel
-    : PUCE BACKTICK TEXT BACKTICK date COMMA MOT_JE MOT_ACHETER TEXT COMMA MOT_VALANT argent COMMA MATERIEL_APPRECIATION MOT_ANNUELLEMENT_DE ENTIER PERCENT COMMA MOT_DEPUIS TEXT
+    : PUCE BACKTICK TEXT BACKTICK date COMMA MOT_JE? MOT_ACHETER TEXT COMMA MOT_VALANT argent COMMA MATERIEL_APPRECIATION MOT_ANNUELLEMENT_DE ENTIER PERCENT COMMA MOT_DEPUIS TEXT
     ;
 possedeMateriel
-    : PUCE BACKTICK TEXT BACKTICK date COMMA MOT_JE MOT_POSSEDER TEXT COMMA MOT_VALANT argent COMMA MATERIEL_APPRECIATION MOT_ANNUELLEMENT_DE ENTIER PERCENT
+    : PUCE BACKTICK TEXT BACKTICK date COMMA MOT_JE? MOT_POSSEDER TEXT COMMA MOT_VALANT argent COMMA MATERIEL_APPRECIATION MOT_ANNUELLEMENT_DE ENTIER PERCENT
     ;
 
 /* Commun */
