+++ src/main/java/school/hei/patrimoine/patrilang/modele/MaterielAppreciationType.java
@@ -1,4 +1,4 @@
-package school.hei.patrimoine.patrilang.model;
+package school.hei.patrimoine.patrilang.modele;
 
 import lombok.Getter;
