+++ src/main/java/school/hei/patrimoine/patrilang/visitors/BaseVisitor.java
@@ -9,7 +9,7 @@ import org.antlr.v4.runtime.tree.TerminalNode;
 import school.hei.patrimoine.modele.Argent;
 import school.hei.patrimoine.modele.Devise;
 import school.hei.patrimoine.patrilang.antlr.PatriLangParser;
-import school.hei.patrimoine.patrilang.model.MaterielAppreciationType;
+import school.hei.patrimoine.patrilang.modele.MaterielAppreciationType;
 
 public class BaseVisitor {
   public static Argent visitArgent(PatriLangParser.ArgentContext ctx) {
