+++ src/main/java/school/hei/patrimoine/patrilang/visitors/TransferArgentVisitor.java
@@ -4,16 +4,16 @@ import static school.hei.patrimoine.patrilang.antlr.PatriLangParser.FluxArgentTr
 import static school.hei.patrimoine.patrilang.visitors.BaseVisitor.*;
 
 import java.time.LocalDate;
-import java.util.function.Function;
 import lombok.RequiredArgsConstructor;
 import school.hei.patrimoine.modele.Argent;
 import school.hei.patrimoine.modele.possession.Compte;
 import school.hei.patrimoine.modele.possession.TransfertArgent;
+import school.hei.patrimoine.patrilang.modele.PossessionGetter;
 
 @RequiredArgsConstructor
 public class TransferArgentVisitor
     implements PossessionVisitor<TransfertArgent, FluxArgentTransfererContext> {
-  private final Function<String, Compte> compteGetter;
+  private final PossessionGetter<Compte> compteGetter;
 
   @Override
   public TransfertArgent visit(FluxArgentTransfererContext ctx) {
